[JENKINS-63673] Allow configuring suppressed events in NoTriggerBranchProperty
The branch property was by default suppressing all triggers: indexing and events (e.g. webhooks). The indexing mechanism is problematic for cloud environments when Jenkins configuration is not kept on a persistent volume. Every Jenkins creation schedules builds of all branches and pull requests which may block the instance for many hours or even days. This change introduces a configuration option which allows administrators/developers to decide if both or only one trigger's kind should be suppressed. It is useful for the mentioned cloud environments because the indexing mechanism could be disabled without ignoring webhooks.